backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
openai-whisper>=20231117
yt-dlp>=2024.01.01
aiofiles>=23.2.0
